Expose mock DynamoDB query stream
This is particularly useful for testing.
